serving_onnx/fastapi_onnx/requirements.txt
fastapi
uvicorn
numpy
pillow
torch
torchvision
onnx
onnxruntime-gpu
opencv-python-headless
pydantic